requirements.txt
fastapi==0.115.12
pydantic==2.11.4
python-dotenv==1.1.0
SQLAlchemy==2.0.40
uvicorn==0.34.2
psycopg2
